// src/screens/FlashLeads/interface/item.interface.ts
export interface ILead {
    lead_appoinments_fixed_date: string;
    painting_type: string;
    lead_guid: string;
    lead_id: string;
    lead_contact_name: string;
    lead_address: string;
    lead_contact_no: string;
    lead_date: string;
    lead_time: string;
    lead_status: string;
    lead_type: string;
    painter_allocated_date: string;
    allocated_pending_time: string;
}